+++ src/utils.js
@@ -169,7 +169,7 @@ export const escapeValue = (val) => {
   }
 
   return escape(val);
-}
+};
 
 export function buildTemplate(template = '') {
   return dataSource => {